+++ drake/examples/kuka_iiwa_arm/iiwa_simulation.cc
@@ -1,7 +1,7 @@
 #include "iiwa_simulation.h"
 
-#include "drake/systems/LCMSystem.h"
 #include "drake/common/drake_path.h"
+#include "drake/systems/LCMSystem.h"
 
 namespace drake {
 namespace examples {
@@ -14,7 +14,7 @@ std::shared_ptr<RigidBodySystem> CreateKukaIiwaSystem(void) {
   auto rigid_body_system = std::allocate_shared<RigidBodySystem>(
       Eigen::aligned_allocator<RigidBodySystem>());
 
-  rigid_body_system->addRobotFromFile(
+  rigid_body_system->AddModelInstanceFromFile(
       drake::GetDrakePath() + "/examples/kuka_iiwa_arm/urdf/iiwa14.urdf",
       DrakeJoint::FIXED);
 
@@ -47,16 +47,88 @@ std::shared_ptr<RigidBodySystem> CreateKukaIiwaSystem(void) {
   return rigid_body_system;
 }
 
-drake::SimulationOptions SetupSimulation(void) {
+drake::SimulationOptions SetupSimulation(double initial_step_size) {
   // Specifies the simulation options.
   drake::SimulationOptions options;
   options.realtime_factor = 0;  // As fast as possible.
-  options.initial_step_size = 0.002;
+  options.initial_step_size = initial_step_size;
 
   // Prevents exception from being thrown when simulation runs slower than real
   // time, which it most likely will given the small step size.
   options.warn_real_time_violation = true;
   return (options);
+}
+
+void CheckLimitViolations(
+    const std::shared_ptr<drake::RigidBodySystem> rigid_body_sys,
+    const Eigen::VectorXd& final_robot_state) {
+  const auto& tree = rigid_body_sys->getRigidBodyTree();
+  int num_positions = rigid_body_sys->number_of_positions();
+  int num_velocities = rigid_body_sys->number_of_velocities();
+
+  // Ensures the size of the output is correct.
+  if (final_robot_state.size() !=
+      static_cast<int>(rigid_body_sys->getNumOutputs())) {
+    throw std::runtime_error(
+        "ERROR: Size of final robot state (" +
+        std::to_string(final_robot_state.size()) +
+        ") does not match size of rigid body system's output (" +
+        std::to_string(rigid_body_sys->getNumOutputs()) + ").");
+  }
+
+  // Ensures the number of position states equals the number of velocity states.
+  if (num_positions != num_velocities) {
+    throw std::runtime_error("ERROR: Number of positions (" +
+                             std::to_string(num_positions) +
+                             ") does not match the number of velocities (" +
+                             std::to_string(num_velocities) + ").");
+  }
+
+  // Ensures the number of position and velocity states match the size of the
+  // final robot state.
+  if ((num_positions + num_velocities) != final_robot_state.size()) {
+    throw std::runtime_error(
+        "ERROR: Total number of positions and velocities (" +
+        std::to_string(num_positions + num_velocities) +
+        ") does not match size of robot state (" +
+        std::to_string(final_robot_state.size()) + ").");
+  }
+
+  // Ensures the robot's joints are within their position limits.
+  std::vector<std::unique_ptr<RigidBody>>& bodies = tree->bodies;
+  for (int robot_state_index = 0, body_index = 0;
+       body_index < static_cast<int>(bodies.size()); ++body_index) {
+    // Skips rigid bodies without a parent (this includes the world link).
+    if (!bodies[body_index]->hasParent()) continue;
+
+    const DrakeJoint& joint = bodies[body_index]->getJoint();
+    const Eigen::VectorXd& min_limit = joint.getJointLimitMin();
+    const Eigen::VectorXd& max_limit = joint.getJointLimitMax();
+
+    // Defines a joint limit tolerance. This is the amount in radians over which
+    // joint position limits can be violated and still be considered within the
+    // limits. Once we are able to model joint limits via
+    // constraints, we may be able to remove the need for this tolerance value.
+    const double kJointLimitTolerance = 0.0261799;  // 1.5 degrees.
+
+    for (int ii = 0; ii < joint.getNumPositions(); ++ii) {
+      double position = final_robot_state[robot_state_index++];
+      if (position < min_limit[ii] - kJointLimitTolerance) {
+        throw std::runtime_error("ERROR: Joint " + joint.getName() + " (DOF " +
+                                 joint.getPositionName(ii) +
+                                 ") violated minimum position limit (" +
+                                 std::to_string(position) + " < " +
+                                 std::to_string(min_limit[ii]) + ").");
+      }
+      if (position > max_limit[ii] + kJointLimitTolerance) {
+        throw std::runtime_error("ERROR: Joint " + joint.getName() + " (DOF " +
+                                 joint.getPositionName(ii) +
+                                 ") violated maximum position limit (" +
+                                 std::to_string(position) + " > " +
+                                 std::to_string(max_limit[ii]) + ").");
+      }
+    }
+  }
 }
 
 }  // namespace kuka_iiwa_arm