+++ drake/systems/framework/primitives/test/adder_test.cc
@@ -85,6 +85,11 @@ TEST_F(AdderTest, AdderIsStateless) {
   EXPECT_EQ(nullptr, context_->get_state().continuous_state);
 }
 
+// Asserts that adders are systems with direct feedthrough inputs.
+TEST_F(AdderTest, AdderIsDirectFeedthrough) {
+  EXPECT_TRUE(adder_->has_any_direct_feedthrough());
+}
+
 }  // namespace
 }  // namespace systems
 }  // namespace drake