+++ drake/util/test/testQuatmex.cpp
@@ -3,6 +3,7 @@
 #include <tuple>
 
 #include "drake/math/autodiff.h"
+#include "drake/math/autodiff_gradient.h"
 #include "drake/math/quaternion.h"
 #include "drake/util/drakeGeometryUtil.h"
 #include "drake/core/Gradient.h"