+++ CHANGELOG.md
@@ -23,7 +23,9 @@ Unreleased: changes on master, not yet released
 [//]: # "Altered functionality or APIs."
 ### Changed
 
+ - [#3115][] Modified SDF parser method names to be style guide compliant and more meaningful.
  - [#3078][] Changed `RigidBodyTree::kWorldLinkName` to be `RigidBodyTree::kWorldName`.
+ - [#3056][] Renamed methods that add model instances to `RigidBodyTree` and `RigidBodySystem`.
  - [#3049][] Changed `AddRobotFromURDF*` to be `AddModelInstanceFromURDF*`.
  - [#3003][] Made `RigidBodyFrame` member variables private. Added accessors.
  - [#3010][] All header file names under `solvers` are now spelled with lower case and underscore names.
@@ -124,4 +126,6 @@ Changes in version v0.9.11 and before are not provided.
 [#3003]: https://github.com/RobotLocomotion/drake/issues/3003
 [#3010]: https://github.com/RobotLocomotion/drake/issues/3010
 [#3049]: https://github.com/RobotLocomotion/drake/issues/3049
+[#3056]: https://github.com/RobotLocomotion/drake/issues/3056
 [#3078]: https://github.com/RobotLocomotion/drake/issues/3078
+[#3115]: https://github.com/RobotLocomotion/drake/issues/3115